autopush: optimize away ivar usage under MRI
We know that all versions of MRI have a small RFile structure
that is allocated in the same object slots as other Ruby types
and also zeroed on allocation.

This optimization enables us to fall back to using ivars in
case MRI changes or if we're used on other Rubies.